foundry-agent/requirements.txt
azure-ai-projects>=1.0.0b7
azure-identity>=1.15.0
